specialties: service impl

+++ src/main/java/com/hieunt/medlink/app/services/hospital/SpecialtiesServiceImpl.java
@@ -5,6 +5,7 @@ import com.hieunt.medlink.app.mappers.SpecialtiesMapper;
 import com.hieunt.medlink.app.repositories.SpecialtiesRepository;
 import com.hieunt.medlink.app.requests.hospital.SpecialtyRequest;
 import com.hieunt.medlink.app.responses.BaseResponse;
+import com.hieunt.medlink.pkg.error.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -12,63 +13,56 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class SpecialtiesServiceImpl implements SpecialtiesService{
+public class SpecialtiesServiceImpl implements SpecialtiesService {
     private final SpecialtiesRepository specialtiesRepository;
     private final SpecialtiesMapper specialtiesMapper;
 
     @Override
     public BaseResponse<SpecialtiesEntity> createSpecialty(SpecialtyRequest specialty) {
-        try {
-            SpecialtiesEntity specialtiesEntity = specialtiesMapper.toEntity(specialty);
+        SpecialtiesEntity specialtiesEntity = specialtiesMapper.toEntity(specialty);
+        if (specialtiesEntity != null) {
             specialtiesRepository.save(specialtiesEntity);
-            return new BaseResponse<>("creating specialty successfully", specialtiesEntity);
-        } catch (Exception e) {
-            throw new RuntimeException("error creating specialty: " + e.getMessage(), e);
         }
+
+        return new BaseResponse<>("creating specialty successfully", specialtiesEntity);
     }
 
     @Override
     public BaseResponse<SpecialtiesEntity> updateSpecialty(Long id, SpecialtyRequest specialty) {
-        try {
-            SpecialtiesEntity specialtiesEntity = getSpecialtyById(id).getData();
+        SpecialtiesEntity specialtiesEntity = getSpecialtyById(id).getData();
+
+        if (specialtiesEntity != null) {
             specialtiesMapper.toEntity(specialty, specialtiesEntity);
             specialtiesRepository.save(specialtiesEntity);
             return new BaseResponse<>("updating specialty successfully", specialtiesEntity);
-        } catch (Exception e) {
-            throw new RuntimeException("error updating specialty: " + e.getMessage(), e);
         }
+
+        throw new ResourceNotFoundException("Specialty not found");
     }
 
     @Override
     public BaseResponse<SpecialtiesEntity> getSpecialtyById(Long id) {
-        try {
-            SpecialtiesEntity specialtiesEntity = specialtiesRepository.findById(id)
-                    .orElseThrow(() -> new RuntimeException("Specialty with id " + id + " not found"));
-            return new BaseResponse<>("getting specialty successfully", specialtiesEntity);
-        } catch (Exception e) {
-            throw new RuntimeException("error getting specialty: " + e.getMessage(), e);
-        }
+        SpecialtiesEntity specialtiesEntity = specialtiesRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Specialty not found"));
+
+        return new BaseResponse<>("getting specialty successfully", specialtiesEntity);
     }
 
     @Override
     public BaseResponse<SpecialtiesEntity> deleteSpecialty(Long id) {
-        try {
-            SpecialtiesEntity specialtiesEntity = getSpecialtyById(id).getData();
+        SpecialtiesEntity specialtiesEntity = getSpecialtyById(id).getData();
+
+        if (specialtiesEntity != null) {
             specialtiesRepository.delete(specialtiesEntity);
             return new BaseResponse<>("deleting specialty successfully", specialtiesEntity);
-        } catch (Exception e) {
-            throw new RuntimeException("error deleting specialty: " + e.getMessage(), e);
         }
+
+        throw new ResourceNotFoundException("Specialty not found");
     }
 
     @Override
     public BaseResponse<Page<SpecialtiesEntity>> filterSpecialties(Long hospitalId, Pageable pageable) {
-        try {
-            // Implement filtering logic here
-            Page<SpecialtiesEntity> specialtiesEntities = specialtiesRepository.filterSpecialties(hospitalId, pageable);
-            return new BaseResponse<>("filtering specialties successfully", specialtiesEntities);
-        } catch (Exception e) {
-            throw new RuntimeException("error filtering specialties: " + e.getMessage(), e);
-        }
+        Page<SpecialtiesEntity> specialtiesEntities = specialtiesRepository.filterSpecialties(hospitalId, pageable);
+        return new BaseResponse<>("filtering specialties successfully", specialtiesEntities);
     }
 }
